+++ cocoatouch/src/main/java/org/robovm/cocoatouch/uikit/UIKeyboardAppearance.java
@@ -21,8 +21,8 @@ public enum UIKeyboardAppearance implements ValuedEnum {
     Default(0),
     Alert(1);
 
-    private final int n;
+    private final long n;
 
-    private UIKeyboardAppearance(int n) { this.n = n; }
-    public int value() { return n; }
+    private UIKeyboardAppearance(long n) { this.n = n; }
+    public long value() { return n; }
 }
